pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-assistant-hub"
version = "0.1.0"
description = "Modular MCP server aggregating external service tools"
authors = [{ name = "AI Assistant", email = "developer@example.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "mcp>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[project.scripts]
ai-assistant-hub = "ai_assistant_hub.server.main:main"
